requirements.txt
# Core
Django==4.0.4
gunicorn==20.1.0

# For the persistence stores
psycopg2-binary==2.9.3

# Rest apis
djangorestframework==3.13.1
dj-rest-auth[with_social]==2.2.4
django-cors-headers==3.12.0
djangorestframework-simplejwt==5.1.0
djangorestframework-camel-case==1.3.0
django-filter==21.1

# Developer Tools
flake8==4.0.1
drf-yasg==1.20.0

# Testing
mock==4.0.3
factory-boy==3.2.1
